move trusted device fingerprint configuration to separate configuration class
